Use open_memmap to create file instead of save + load
Previously, we would first save the NumPy file just to open it
immediately after. This caused us to write the NumPy header and then
read it back immediately after.

This commit instead uses open_memmap directly to create and load the
file in the same operation. The data is written immediately after.