Add types to top-level import to simplify porting from h5py
Previously, types like Dataset would have to be imported
through exdir.core.Dataset, which exposes an internal implementation
detail (everything being in exdir.core).

Type-checking by comparing the type to the one imported is common in h5py code,
which means porting also is cumbersome whenever a comparison is performed.
This changes fixes this issues by exposing the types in the root
import.