Add plugin data objects and give plugins access to metadata (#42)
This change introduces data objects for attribute plugins and dataset plugins. This makes it easier to extend plugins with more data in the future, as plugins only have to rely on the currently available information in a data object.

Further, this change gives plugins access to all the metadata in the exdir.yaml file. This gives great power to plugins, to the level of even being able to change the exdir verison (with great power comes great responsibility).

The quantities plugin and the tests have been updated to reflect these changes.

Co-authored-by: Milad Hobbi Mobarhan <m@milad.no>