Remove support for Python 3.8, introduce 3.11
The latest versions of NumPy dropped support for Python 3.8. Exdir
follows the support matrix of NumPy to avoid being stuck on an old
version of NumPy.

Unfortunately, I was required to pin the version of ruamel.yaml to get
the build to work in CI. Otherwise it would fail with the following
error:

```
AttributeError: module 'ruamel.yaml' has no attribute 'YAML'
```

I will hopefully be unpin it in the future.